Split iOS progress series merge tests
